refactor: move forgot/reset password pages to root, source expires-in from config
Address the last two open review items from PR #119:

- #9: '1 hour' was hard-coded in the ForgotPassword success view. Now
  sourced from TickfloConfig.PasswordResetTokenMaxAgeSeconds via a
  FormatExpiresIn helper on the page model (same formatting rules as
  the email template).
- #11: Pages moved from Pages/Account/ to Pages/ following the
  SetPassword precedent. Routes are now /forgot-password and
  /reset-password. Updated the email reset link, the Login link, and
  the corresponding test assertion to match.

No database changes. 69/69 tests pass.
