+++ Tickflo.Web/Pages/ForgotPassword.cshtml
@@ -1,5 +1,5 @@
 @page
-@model Tickflo.Web.Pages.Account.ForgotPasswordModel
+@model Tickflo.Web.Pages.ForgotPasswordModel
 @{
     ViewData["Title"] = "Forgot password";
 }
@@ -17,7 +17,7 @@
                     </svg>
                     <span>If an account exists for that email, we just sent a password reset link. Check your inbox.</span>
                 </div>
-                <p class="text-sm text-slate-400 mt-4">The link will expire in 1 hour. You can close this tab.</p>
+                <p class="text-sm text-slate-400 mt-4">The link will expire in @Model.ExpiresIn. You can close this tab.</p>
             }
             else
             {
