+++ Tickflo.Web/Pages/ForgotPassword.cshtml.cs
@@ -1,12 +1,16 @@
-namespace Tickflo.Web.Pages.Account;
+namespace Tickflo.Web.Pages;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Tickflo.Core.Config;
 using Tickflo.Core.Services.Authentication;
 
-public class ForgotPasswordModel(IPasswordResetRequestService passwordResetRequestService) : PageModel
+public class ForgotPasswordModel(
+    IPasswordResetRequestService passwordResetRequestService,
+    TickfloConfig tickfloConfig) : PageModel
 {
     private readonly IPasswordResetRequestService passwordResetRequestService = passwordResetRequestService;
+    private readonly TickfloConfig tickfloConfig = tickfloConfig;
 
     [BindProperty]
     public string Email { get; set; } = string.Empty;
@@ -14,6 +18,8 @@ public class ForgotPasswordModel(IPasswordResetRequestService passwordResetReque
     public bool Submitted { get; private set; }
 
     public string? ErrorMessage { get; private set; }
+
+    public string ExpiresIn => FormatExpiresIn(this.tickfloConfig.PasswordResetTokenMaxAgeSeconds);
 
     public void OnGet()
     {
@@ -38,5 +44,22 @@ public class ForgotPasswordModel(IPasswordResetRequestService passwordResetReque
             this.ErrorMessage = ex.Message;
             return this.Page();
         }
+    }
+
+    private static string FormatExpiresIn(int maxAgeInSeconds)
+    {
+        if (maxAgeInSeconds % 3600 == 0)
+        {
+            var hours = maxAgeInSeconds / 3600;
+            return hours == 1 ? "1 hour" : $"{hours} hours";
+        }
+
+        if (maxAgeInSeconds % 60 == 0)
+        {
+            var minutes = maxAgeInSeconds / 60;
+            return minutes == 1 ? "1 minute" : $"{minutes} minutes";
+        }
+
+        return $"{maxAgeInSeconds} seconds";
     }
 }
