+++ Tickflo.Web/Pages/Login.cshtml
@@ -22,7 +22,7 @@
                     <input id="password" asp-for="Input.Password" type="password" class="input bg-neutral-800/50 border-white/10 text-slate-100 focus:border-primary h-12 w-full"
                         placeholder="Password" value=""/>
                     <div class="mt-2">
-                        <a class="link link-primary link-hover text-sm" href="/account/forgot-password">
+                        <a class="link link-primary link-hover text-sm" href="/forgot-password">
                             Forgot password?
                         </a>
                     </div>
