+++ Tickflo.Web/Pages/ResetPassword.cshtml
@@ -1,5 +1,5 @@
 @page
-@model Tickflo.Web.Pages.Account.ResetPasswordModel
+@model Tickflo.Web.Pages.ResetPasswordModel
 @{
     ViewData["Title"] = "Reset password";
 }
@@ -13,7 +13,7 @@
             {
                 <p class="text-sm text-slate-400 mb-4">This password reset link is invalid or has expired.</p>
                 @await Html.PartialAsync("_ErrorAlert", "Request a new link below.")
-                <a href="/account/forgot-password" class="btn btn-primary mt-4 w-full rounded-full h-12">
+                <a href="/forgot-password" class="btn btn-primary mt-4 w-full rounded-full h-12">
                     Request a new link
                 </a>
             }
